+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ Changelog
 Next
 ----
 
+2023.05.21
+------------
+
 * Breaking change: the ``vws.exceptions.custom_exceptions.ActiveMatchingTargetsDeleteProcessing`` exception has been removed as Vuforia no longer returns this error.
 
 2023.03.25